Refactor (apps/pet-shop): Move `ps-products.type`
