+++ apps/pet-shop/src/app/integrator/ui/ps-cart/ps-cart.component.ts
@@ -18,8 +18,8 @@ import {
 } from '@angular/material/table'
 
 // # Internal modules
-import { type ProductTableViewModel } from '../../../shop/products/shared/product-table-view.model'
 import { PSEmptinessComponent } from '../../ui-kit/ps-emptiness/ps-emptiness.component'
+import type { ProductTableViewModel } from '../ps-products/ps-products.type'
 import { PSCartService } from './ps-cart.service'
 import { type PSCartState } from './ps-cart.service.type'
