+++ apps/pet-shop/src/app/integrator/ui/ps-cart/ps-cart.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@angular/core'
 import { BehaviorSubject, type Observable } from 'rxjs'
 
 // # Internal modules
-import type { ProductTableViewModel } from '../../../shop/products/shared/product-table-view.model'
+import type { ProductTableViewModel } from '../ps-products/ps-products.type'
 import type { PSCartState } from './ps-cart.service.type'
 
 @Injectable({
