+++ apps/pet-shop/src/app/integrator/ui/ps-cart/ps-cart.service.type.ts
@@ -1,4 +1,4 @@
-import type { ProductTableViewModel } from '../../../shop/products/shared/product-table-view.model'
+import type { ProductTableViewModel } from '../ps-products/ps-products.type'
 
 export interface PSCartState {
   readonly items: readonly ProductTableViewModel[]
